Refactor producer-consumer impl of challenge problem
