+++ demos/challenge/CMakeLists.txt
@@ -13,11 +13,11 @@ set(COMMON_SRC base64.c tweetnacl.c )
 add_executable(challenge_baseline challenge_baseline.c ${COMMON_SRC})
 target_compile_options(challenge_baseline PRIVATE ${BUILD_FLAGS})
 
-add_executable(challenge_socket_main challenge_socket_main.c ${COMMON_SRC})
+add_executable(challenge_socket_main challenge_socket_main.c challenge_socket_common.c ${COMMON_SRC})
 target_compile_options(challenge_socket_main PRIVATE ${BUILD_FLAGS})
-add_executable(challenge_socket_encrypt1 challenge_socket_encrypt1.c ${COMMON_SRC})
+add_executable(challenge_socket_encrypt1 challenge_socket_encrypt1.c challenge_socket_common.c ${COMMON_SRC})
 target_compile_options(challenge_socket_encrypt1 PRIVATE ${BUILD_FLAGS})
-add_executable(challenge_socket_encrypt2 challenge_socket_encrypt2.c ${COMMON_SRC})
+add_executable(challenge_socket_encrypt2 challenge_socket_encrypt2.c challenge_socket_common.c ${COMMON_SRC})
 target_compile_options(challenge_socket_encrypt2 PRIVATE ${BUILD_FLAGS})
 
 add_executable(challenge_multithreaded challenge_multithreaded.c ${COMMON_SRC})
