+++ demos/challenge/challenge_socket_main.c
@@ -10,6 +10,7 @@
 
 #include "tweetnacl.h"
 #include "base64.h"
+#include "challenge_socket_common.h"
 
 #define KEY_BYTES crypto_secretbox_KEYBYTES
 #define NONCE_BYTES crypto_secretbox_NONCEBYTES
@@ -20,109 +21,6 @@
 #if DELTA_ZERO_BYTES < 0
 #error "crypto_secretbox_ZEROBYTES is assumed to be >= crypto_secretbox_BOXZEROBYTES"
 #endif
-
-int reader_open(int port) {
-    int err, rv;
-    int fd, server_fd;
-    struct sockaddr_in addr;
-    struct linger lo;
-
-    server_fd = socket(AF_INET, SOCK_STREAM, 0);
-    if (server_fd < 0) {
-        return server_fd;
-    }
-
-    memset(&addr, 0, sizeof(struct sockaddr_in));
-    addr.sin_family = AF_INET;
-    addr.sin_addr.s_addr = inet_addr("127.0.0.1");
-    addr.sin_port = htons(port);
-
-    int enable = 1;
-    rv = setsockopt(server_fd, SOL_SOCKET, SO_REUSEADDR, &enable, sizeof(int));
-    if (rv < 0) {
-        err = errno;
-        close(server_fd);
-        errno = err;
-        return rv;
-    }
-
-    rv = bind(server_fd, (struct sockaddr *)&addr, sizeof(struct sockaddr_in));
-    if (rv < 0) {
-        err = errno;
-        close(server_fd);
-        errno = err;
-        return rv;
-    }
-
-    rv = listen(server_fd, 0);
-    if (rv < 0) {
-        err = errno;
-        close(server_fd);
-        errno = err;
-        return rv;
-    }
-
-    fd = accept(server_fd, NULL, NULL);
-
-    if (fd < 0) {
-        err = errno;
-        close(server_fd);
-        errno = err;
-        return fd;
-    }
-
-    lo.l_onoff = 1;
-    lo.l_linger = 0;
-    rv = setsockopt(fd, SOL_SOCKET, SO_LINGER, &lo, sizeof(lo));
-    if (rv < 0) {
-        err = errno;
-        close(fd);
-        close(server_fd);
-        errno = err;
-        return rv;
-    }
-
-    close(server_fd);
-    return fd;
-}
-
-// param port is the port of the reader
-int writer_open(int port) {
-    int err, rv, fd;
-    struct sockaddr_in addr;
-
-    fd = socket(AF_INET, SOCK_STREAM, 0);
-    if (fd < 0) {
-        return fd;
-    }
-
-    addr.sin_family = AF_INET;
-    addr.sin_addr.s_addr = inet_addr("127.0.0.1");
-    addr.sin_port = htons(port);
-
-    for (;;) {
-        err = errno;
-        rv = connect(fd, (struct sockaddr *)&addr, sizeof(addr));
-        if (rv < 0) {
-            if ((errno == ENOENT) || (errno == ECONNREFUSED)) {
-                struct timespec req;
-                errno = err;
-                req.tv_sec = 0;
-                req.tv_nsec = 1e8;
-                rv = nanosleep(&req, NULL);
-                if (rv == 0) {
-                    continue;
-                }
-            }
-            err = errno;
-            close(fd);
-            errno = err;
-            return rv;
-        }
-        return fd;
-    }
-    return -1;
-}
 
 int main() {
     char input[80 + DELTA_ZERO_BYTES] = {0};
